Smart-Agriculture-Advisory-System/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.1
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
shap>=0.43.0
joblib==1.3.2
redis==5.0.1
celery==5.3.4
twilio==8.10.0
aiohttp==3.9.1
python-multipart==0.0.6
cryptography==41.0.7
